hot fix

+++ 0x15-file_io/1-create_file.c
@@ -8,9 +8,9 @@
  *
  * Return: int length
  */
-size_t _strlen(char *str)
+int  _strlen(char *str)
 {
-	size_t i = 0;
+	int i = 0;
 
 	for (i = 0; str[i] != '\0'; i++)
 		;
